backend/requirements.txt
asgiref==3.8.1
Django==5.0.4
idna==3.7
django-cors-headers==4.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
pillow==10.3.0
psycopg2==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.6.0
redis==5.0.4
requests==2.31.0
django-filter==24.2
django-redis==5.4.0
django-mpesa==2.0.14
stripe==9.8.0
python-decouple==3.8